pyproject.toml
[tool.poetry]
name = "python-movie-api-example"
version = "0.1.0"
description = ""
authors = ["Carlos Kidman <carlos@qap.dev>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
requests = "^2.26.0"
pydantic = "^1.8.2"
python-dotenv = "^0.19.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pytest-cov = "^2.12.1"
black = "^21.8b0"
flake8 = "^3.9.2"
poethepoet = "^0.10.0"

[tool.black]
line-length = 120

[tool.poe.tasks]
test = { "cmd" = "pytest tests --cov=. --cov-report term-missing", "help" = "Run Tests and get Code Coverage Report" }
lint = [{ "ref" = "_lint-syntax" }, { "ref" = "_lint-full" }]

# private tasks
_lint-syntax = { "cmd" = "flake8 ./tmdb ./tests --select=E9,F63,F7,F82 --show-source", "help" = "Lint for syntax error and undefined names" }
_lint-full = { "cmd" = "flake8 ./tmdb ./tests --exit-zero --max-complexity=10", "help" = "Lint for style, complexity, and everything else" }

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
